% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_diarias.R
\name{casos_diarios}
\alias{casos_diarios}
\title{casos_diarios}
\usage{
casos_diarios(data, DEPARTAMENTO = NULL, mediamovil = NA)
}
\arguments{
\item{data}{Data frame creado con la función da_positivos()}

\item{DEPARTAMENTO}{Departamento del Perú. Callao está listado como departamento}

\item{mediamovil}{Integer para calcular o graficar la media móvil. Eje: 1,2,3.}
}
\value{

}
\description{
casos_diarios
}
