% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidPeru-package.R
\docType{package}
\name{covidPeru-package}
\alias{covidPeru}
\alias{covidPeru-package}
\title{covidPeru: Funciones para procesar información sobre COVID-19 en Perú}
\description{
Paquete para analizar la información de COVID generada por MINSA.\Todos los datos se extraen de los repositorios de Datos Abiertos.
}
\author{
\strong{Maintainer}: José Incio \email{jincio@gmail.com} (\href{https://orcid.org/0000-0001-6866-0644}{ORCID})

}
\keyword{internal}
