% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{da_fallecidos}
\alias{da_fallecidos}
\title{load fallecidos Covid}
\usage{
da_fallecidos()
}
\value{
Data.frame con la información de "fallecidos" (datos abiertos). Agregando la variable fecha y semana.
}
\description{
Carga la base de datos de Fallecidos desde la página de datos abiertos
}
\examples{

}
