% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{da_positivos}
\alias{da_positivos}
\title{Load positive COVID cases}
\usage{
da_positivos()
}
\value{
data_frame
}
\description{
Cargal la base de datos de casos confirmados desde la página de datos abiertos
}
