% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceso_muertos.R
\name{exceso_muertes}
\alias{exceso_muertes}
\title{exceso_muertes}
\usage{
exceso_muertes(data, DEPARTAMENTO = NULL, metodo = "M2020")
}
\arguments{
\item{metodo}{default="Mactual". Mactual considera el promedio de las primeras 11 semanas del 2020. "Mhistorico" compara la semana del 2020 con las el promedio de fallecidos en la misma semana en en el 2017, 2018,2019.}
}
\value{

}
\description{
exceso_muertes
}
