% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_diarias.R
\name{fallecidos_crecimiento}
\alias{fallecidos_crecimiento}
\title{Fallecidos_ratio}
\usage{
fallecidos_crecimiento(
  data,
  DEPARTAMENTO = NULL,
  mediamovil = NA,
  semanal = FALSE
)
}
\arguments{
\item{data}{Data frame creado con la función da_fallecidos()}

\item{DEPARTAMENTO}{Departamento del Perú. Callao está listado como departamento}

\item{mediamovil}{Integer para calcular o graficar la media móvil. Eje: 1,2,3.}

\item{semanal}{Boleano que permite obtener la información en frecuencia semanal}
}
\value{

}
\description{
Fallecidos_ratio
}
