% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_diarias.R
\name{fallecidos_diarios}
\alias{fallecidos_diarios}
\title{fallecidos_diarios}
\usage{
fallecidos_diarios(data, DEPARTAMENTO = NULL, mediamovil = NA)
}
\arguments{
\item{data}{Data frame creado con la función da_fallecidos()}

\item{DEPARTAMENTO}{Departamento del Perú. Callao está listado como departamento}

\item{mediamovil}{Integer para calcular o graficar la media móvil. Eje: 1,2,3.}
}
\value{

}
\description{
fallecidos_diarios
}
