% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_exceso.R
\name{panel_exceso}
\alias{panel_exceso}
\title{panel exceso}
\usage{
panel_exceso(data)
}
\arguments{
\item{data}{a un data frame creado con da_sinadef()}
}
\value{
a ggplot object
}
\description{
\code{panel_exceso} es una funcion creada por Gabriel Carrasco
}
