% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficos.R
\name{piramide_fcovid}
\alias{piramide_fcovid}
\title{piramide_fcovid}
\usage{
piramide_fcovid(data, DEPARTAMENTO = NULL)
}
\arguments{
\item{data}{un data frame con \code{da_fallecidos()}}

\item{DEPARTAMENTO}{}
}
\value{

}
\description{
\code{piramide_fcovid} es una funcion creada por Jesus Gamboa (twitter:jgamboa)
}
