% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficos.R
\name{piramide_pcovid}
\alias{piramide_pcovid}
\title{piramide positivis}
\usage{
piramide_pcovid(data, DEPARTAMENTO = NULL)
}
\arguments{
\item{data}{a data frame creado por da_positivos()}

\item{DEPARTAMENTO}{}
}
\value{

}
\description{
\code{piramide_fcovid} es una funcion creada por Jesus Gamboa (twitter:jgamboa)
}
