% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_diarias.R
\name{sinadef_diarios}
\alias{sinadef_diarios}
\title{sinadef_diarios}
\usage{
sinadef_diarios(data, DEPARTAMENTO = NULL, mediamovil = NA)
}
\arguments{
\item{data}{Data frame creado con la función da_sinadef()}

\item{DEPARTAMENTO}{Departamento del Perú. Callao está listado como departamento}

\item{mediamovil}{Integer para calcular o graficar la media móvil. Eje: 1,2,3.}
}
\value{

}
\description{
sinadef_diarios
}
