% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SirdModels.R
\name{sird_basico}
\alias{sird_basico}
\title{sird_basico}
\usage{
sird_basico(delta, gamma, theta, N, d)
}
\arguments{
\item{d}{}
}
\value{

}
\description{
\code{panel_exceso} es una funcion creada por Cesar Urquizo (twitter:CUrquizoU)
}
