% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficos.R
\name{sird_graficos}
\alias{sird_graficos}
\title{SIRD graficos}
\usage{
sird_graficos(resultados)
}
\arguments{
\item{resultados}{output generado por la funcion sird_villaverde()}
}
\value{

}
\description{
SIRD graficos
}
