% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SirdModels.R
\name{sird_villaverde}
\alias{sird_villaverde}
\title{Sird_Villaverde}
\usage{
sird_villaverde(
  data,
  DEPARTAMENTO = NULL,
  delta = 0.01,
  gamma = 0.1,
  theta = 0.2
)
}
\arguments{
\item{theta}{}
}
\value{

}
\description{
\code{panel_exceso} es una funcion creada por Cesar Urquizo (twitter:CUrquizoU)
}
