% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_diarias.R
\name{vacunados_diarios}
\alias{vacunados_diarios}
\title{vacunados_diarios}
\usage{
vacunados_diarios(data, DEPARTAMENTO = NULL, mediamovil = NA, semanal = FALSE)
}
\arguments{
\item{data}{Data frame creado con la función da_vacunados()}

\item{DEPARTAMENTO}{Departamento del Perú. Callao está listado como departamento}

\item{mediamovil}{Integer para calcular o graficar la media móvil. Eje: 1,2,3.}

\item{semanal}{Boleano True permite obtener la informacion en frecuencia semanal}
}
\value{

}
\description{
vacunados_diarios
}
