% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_diarias.R
\name{casos_crecimiento}
\alias{casos_crecimiento}
\title{Casos_crecimiento}
\usage{
casos_crecimiento(data, DEPARTAMENTO = NULL, mediamovil = NA, semanal = FALSE)
}
\arguments{
\item{data}{Data frame creado con la función da_positivos()}

\item{DEPARTAMENTO}{Departamento del Perú. Callao está listado como departamento}

\item{mediamovil}{Integer para calcular o graficar la media móvil. Eje: 1,2,3.}

\item{semanal}{Boleano True permite obtener la informacion en frecuencia semanal}
}
\value{

}
\description{
Casos_crecimiento
}
