% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{da_vacunados}
\alias{da_vacunados}
\title{load Vaccinated people in Peru}
\usage{
da_vacunados()
}
\value{
data_frame
}
\description{
Cargal la base de datos de vacunas aplicadas desde la página de datos abiertos
}
