% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceso_muertos.R
\name{exceso_muertes}
\alias{exceso_muertes}
\title{exceso_muertes}
\usage{
exceso_muertes(data, DEPARTAMENTO = NULL, metodo = TRUE)
}
\arguments{
\item{metodo}{Default M2020}
}
\value{

}
\description{
exceso_muertes
}
